@@ -3,45 +3,53 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve -o",
+    "start": "ng serve -o --hmr",
     "build": "ng build",
     "lint": "ng lint",
     "document": "compodoc -p tsconfig.json"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.0.9",
-    "@angular/cdk": "^10.1.3",
-    "@angular/common": "^10.0.9",
-    "@angular/compiler": "^10.0.9",
-    "@angular/core": "^10.0.9",
-    "@angular/flex-layout": "^10.0.0-beta.32",
-    "@angular/forms": "^10.0.9",
-    "@angular/material": "^10.1.3",
-    "@angular/platform-browser": "^10.0.9",
-    "@angular/platform-browser-dynamic": "^10.0.9",
-    "@angular/router": "^10.0.9",
-    "@auth0/angular-jwt": "^5.0.1",
-    "@swimlane/ngx-charts": "^16.0.0",
+    "@angular/animations": "^11.0.0",
+    "@angular/cdk": "^11.0.0",
+    "@angular/common": "^11.0.0",
+    "@angular/compiler": "^11.0.0",
+    "@angular/core": "^11.0.0",
+    "@angular/flex-layout": "^11.0.0-beta.33",
+    "@angular/forms": "^11.0.0",
+    "@angular/material": "^11.0.0",
+    "@angular/platform-browser": "^11.0.0",
+    "@angular/platform-browser-dynamic": "^11.0.0",
+    "@angular/router": "^11.0.0",
+    "@ngx-translate/core": "^13.0.0",
+    "@ngx-translate/http-loader": "^6.0.0",
+    "@auth0/angular-jwt": "^5.0.2",
+    "@ng-bootstrap/ng-bootstrap": "^8.0.0",
     "@popperjs/core": "^2.5.4",
+    "@swimlane/ngx-charts": "^16.0.0",
+    "angular-calendar": "^0.28.22",
+    "angularx-flatpickr": "^6.5.1",
     "bootstrap": "^4.5.3",
+    "date-fns": "^2.16.1",
+    "flatpickr": "^4.6.6",
     "jquery": "^3.5.1",
+    "ng-bootstrap": "^1.6.3",
     "popper.js": "^1.16.1",
     "rxjs": "^6.6.3",
-    "tslib": "^2.0.3",
+    "tslib": "^2.0.0",
     "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1002.0",
-    "@angular/cli": "^10.2.0",
-    "@angular/compiler-cli": "^10.0.9",
-    "@angular/language-service": "^10.2.1",
+    "@angular-devkit/build-angular": "^0.1100.2",
+    "@angular/cli": "^11.0.2",
+    "@angular/compiler-cli": "^11.0.0",
+    "@angular/language-service": "^11.0.2",
     "@compodoc/compodoc": "^1.1.11",
-    "@types/node": "^14.14.6",
-    "codelyzer": "^6.0.1",
+    "@types/node": "^14.14.10",
+    "codelyzer": "^6.0.0",
     "ts-node": "~9.0.0",
     "tslint": "~6.1.0",
-    "typescript": "~3.9.7"
+    "typescript": "~4.0.5"
   },
   "resolutions": {
     "minimist": "^1.2.5"
