@@ -5,9 +5,9 @@
         "test": "cypress run --browser chrome --headless"
     },
     "dependencies": {
-        "typescript": "^4.0.5"
+        "typescript": "^4.1.2"
     },
     "devDependencies": {
-        "cypress": "^5.5.0"
+        "cypress": "^6.0.0"
     }
 }
