@@ -2,7 +2,11 @@ using Microsoft.EntityFrameworkCore;
 using ModelLayer;
 using ModelLayer.Models;
 using RepositoryLayer;
+using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
+using System.Linq;
+using ModelLayer.DataTransferObjects;
 
 namespace ServiceLayer
 {
@@ -18,6 +22,8 @@ namespace ServiceLayer
         Task<string> GetContactInformationAsTextAsync(long id);
 
         Task SendDisclosure(long id);
+
+        Task<ApprovedStatus> ApproveContact(long id);
     }
 
     public class ContactService : ContactRepository, IContactService
@@ -33,20 +39,31 @@ namespace ServiceLayer
         {
             Contact contact = await GetByIdAsync(id);
             string gender = " geehrter Herr ";
+            string genderEn = "Dear Mr ";
             if (contact.Gender == Contact.GenderTypes.FEMALE)
             {
                 gender = " geehrte Frau ";
+                genderEn = "Dear Ms ";
             }
             else if (contact.Gender == Contact.GenderTypes.DIVERS)
             {
                 gender = " geehrt ";
+                genderEn = "Dear Mx ";
             }
 
-            string body = "<h3> Auskunft über gespeicherte Daten </h3> " +
+            string body = "<h5><i> - English version below - </i></h5>" +
+                          "<h3> Auskunft über gespeicherte Daten </h3> " +
                           "<p> Sehr" + gender + contact.Name + ",</p>" +
                           "<p> Sie hatten um eine Auskunft über die zur Ihrer Person in unserem Customer Relationship Management System(CRMS) gespeicherten Daten gebeten. Im angehängten PDF - Dokument erhalten Sie die entsprechende Auskunft gem. Art. 15 EU - DSGVO.</p>" +
                           "<p></p>" +
-                          "<p>Technische Hochschule Nürnberg</p>";
+                          "<p>Technische Hochschule Nürnberg</p>" +
+                          "<h5><i>- English version -</i></h5>" +
+                          "<h3> Information about stored data </h3> " +
+                          "<p>" + genderEn + contact.Name + ",</p>" +
+                          "<p> You asked for information about your personal data stored in our Customer Relationship Management System (CRMS). In the attached PDF - document you will find the corresponding information according to Art. 15 EU - DSGVO.</p>" +
+                          "<p></p>" +
+                          "<p>Nuremberg Institute of Technology</p>";
+
             string pdfBody = await GetContactInformationAsTextAsync(contact.Id);
             pdfBody = pdfBody + "";
             mailProvider.CreateAndSendMail(contact.ContactPossibilities.Mail, "Auskunft über gespeicherte Daten", body,
@@ -79,6 +96,31 @@ namespace ServiceLayer
             }
 
             return await Task.FromResult(contact.ToString());
+        }
+
+        public async Task<ApprovedStatus> ApproveContact(long id) {
+            long maxId = Entities.Max(a => a.Id);
+            if (id > maxId)
+            {
+                return ApprovedStatus.InvalidId; // Invalid ID received!
+            }
+            List<Contact> unapproved = await GetAllUnapprovedContactsWithAllIncludesAsync();
+            foreach (Contact c in unapproved) {
+                if (c.Id == id) {
+                    c.isApproved = true;
+                    bool result = await UpdateAsync(c, id);
+                    if (result)
+                    {
+                        return ApprovedStatus.Approved; // OK
+                    }
+                    else
+                    {
+                        return ApprovedStatus.ErrorInSaving; // failure at trying to save the changes
+                    }
+                }
+            }
+
+            return ApprovedStatus.AlreadyApprovedOrDeleted; // not found as already confirmed or deleted
         }
     }
 }
