@@ -8,6 +8,7 @@ using AutoMapper;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
 using ModelLayer.DataTransferObjects;
 using ModelLayer.Helper;
 using ModelLayer.Models;
@@ -29,6 +30,7 @@ namespace WebApi.Controllers
             this.roleService = roleService;
         }
 
+        [Authorize(Roles = "Einsehen und Überarbeiten des Rollenkonzepts")]
         [HttpGet]
         [SwaggerResponse(HttpStatusCode.OK, typeof(List<RoleDto>), Description = "successfully found")]
         public async Task<IActionResult> Get()
@@ -37,6 +39,7 @@ namespace WebApi.Controllers
             return Ok(roles);
         }
 
+        [Authorize(Roles = "Einsehen und Überarbeiten des Rollenkonzepts")]
         [HttpGet("{id}")]
         [SwaggerResponse(HttpStatusCode.OK, typeof(List<string>), Description = "successfully found")]
         public async Task<IActionResult> GetAllClaims(long id)
@@ -49,6 +52,7 @@ namespace WebApi.Controllers
             return Ok(list);
         }
 
+        [Authorize(Roles = "Anlegen eines Benutzers")]
         [HttpPost]
         [SwaggerResponse(HttpStatusCode.Created, typeof(RoleDto), Description = "successfully created")]
         [SwaggerResponse(HttpStatusCode.BadRequest, typeof(void), Description = "unsuccessfully request")]
@@ -64,7 +68,8 @@ namespace WebApi.Controllers
 
             return BadRequest();
         }
-
+      
+        [Authorize(Roles = "Einsehen und Überarbeiten des Rollenkonzepts")]
         [HttpPut]
         [SwaggerResponse(HttpStatusCode.OK, typeof(void), Description = "successfully updated")]
         [SwaggerResponse(HttpStatusCode.Conflict, typeof(void), Description = "conflict in update process")]
@@ -80,7 +85,8 @@ namespace WebApi.Controllers
                 return Conflict();
             }
         }
-
+  
+        [Authorize(Roles = "Einsehen und Überarbeiten des Rollenkonzepts")]
         [HttpPut("{id}")]
         [SwaggerResponse(HttpStatusCode.OK, typeof(void), Description = "successfully updated")]
         [SwaggerResponse(HttpStatusCode.Conflict, typeof(void), Description = "conflict in update process")]
@@ -96,7 +102,8 @@ namespace WebApi.Controllers
                 return Conflict();
             }
         }
-
+    
+        [Authorize(Roles = "Löschen / Deaktivieren eines Benutzers")]
         [HttpDelete("{id}")]
         [SwaggerResponse(HttpStatusCode.OK, typeof(void), Description = "successfully deleted")]
         [SwaggerResponse(HttpStatusCode.NotFound, typeof(void), Description = "address not found")]
