@@ -3,6 +3,10 @@ import { Routes, RouterModule } from '@angular/router';
 import { HomeComponent } from './home/home.component';
 import { LoginComponent } from './login/login.component';
 import { AuthGuard } from './shared/routing/auth.guard';
+import { OrganizationsListComponent } from './organizations/organizations-list/organizations-list.component';
+import { ApproveContactComponent } from './approve-contact/approve-contact.component';
+import { CalendarComponent } from './calendar/calendar.component';
+
 
 const routes: Routes = [
   {
@@ -27,9 +31,18 @@ const routes: Routes = [
     loadChildren: () => import('./settings/settings.module').then(mod => mod.SettingsModule)
   },
   {
+    path: 'calendar',
+    canActivate: [AuthGuard],
+    component: CalendarComponent
+  },
+  {
     path: 'events',
     canActivate: [AuthGuard],
     loadChildren: () => import('./events/events.module').then(mod => mod.EventsModule)
+  },
+  {
+    path: 'ApproveContacte/:id',
+    component: ApproveContactComponent,
   },
   {
     path: 'statistics',
@@ -39,7 +52,7 @@ const routes: Routes = [
 ];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
+  imports: [RouterModule.forRoot(routes, { relativeLinkResolution: 'legacy' })],
   exports: [RouterModule]
 })
 export class AppRoutingModule { }
