@@ -10,21 +10,22 @@ import { ContactsEditDialogComponent } from './contacts-edit-dialog/contacts-edi
 import { ContactsDetailComponent } from './contacts-detail/contacts-detail.component';
 import { ContactsInfoComponent } from './contacts-info/contacts-info.component';
 import { ContactsDisclosureDialogComponent } from './contacts-disclosure-dialog/contacts-disclosure-dialog.component';
-import { DataProtectionModule } from '../shared/data-protection';
 import {MatSnackBarModule} from '@angular/material/snack-bar';
+import { ContactsSendMailDialogComponent } from './contacts-send-mail-dialog/contacts-send-mail-dialog.component';
 
 @NgModule({
 	imports: [
 		SharedModule,
 		FormsModule,
-		ReactiveFormsModule, MaterialModule, MatDialogModule, ContactsRoutingModule, MatSnackBarModule, DataProtectionModule ],
+		ReactiveFormsModule, MaterialModule, MatDialogModule, ContactsRoutingModule, MatSnackBarModule ],
 	declarations: [
 		ContactsListComponent,
 		ContactsAddDialogComponent,
 		ContactsEditDialogComponent,
 		ContactsDetailComponent,
 		ContactsInfoComponent,
-		ContactsDisclosureDialogComponent
+		ContactsDisclosureDialogComponent,
+		ContactsSendMailDialogComponent
 	]
 })
 export class ContactsModule {}
