@@ -4,6 +4,7 @@ import { MAT_DIALOG_DATA, MatDialogRef, MatDialog } from '@angular/material/dial
 import { FormGroup, FormBuilder } from '@angular/forms';
 import { BaseDialogInput } from '../../shared/form/base-dialog-form/base-dialog.component';
 import { map } from 'rxjs/operators';
+import { PageEvent } from '@angular/material/paginator';
 
 export class ContactOrganizationDtoExtended {
   id: number;
@@ -24,16 +25,21 @@ export class EventsInfoComponent extends BaseDialogInput<EventsInfoComponent> im
   contactsOrganizations: ContactOrganizationDtoExtended[] = new Array<ContactOrganizationDtoExtended>();
   eventsForm: FormGroup;
   dataHistory: ModificationEntryDto[] = new Array<ModificationEntryDto>();
+  modificationsPaginationLength: number;
   columnsContacts = ['wasInvited', 'participated', 'prename', 'name'];
   displayedColumnsDataChangeHistory = ['datum', 'bearbeiter', 'feldname', 'alterWert', 'neuerWert'];
 
   constructor(public dialogRef: MatDialogRef<EventsInfoComponent>,
-              public dialog: MatDialog,
-              @Inject(MAT_DIALOG_DATA) public event: EventDto,
-              private fb: FormBuilder,
-              private modService: ModificationEntryService
+    public dialog: MatDialog,
+    @Inject(MAT_DIALOG_DATA) public event: EventDto,
+    private fb: FormBuilder,
+    private modService: ModificationEntryService
   ) {
     super(dialogRef, dialog);
+    this.dialogRef.backdropClick().subscribe(() => {
+      // Close the dialog
+      dialogRef.close();
+    });
   }
 
   getDate(date: string): string {
@@ -50,50 +56,51 @@ export class EventsInfoComponent extends BaseDialogInput<EventsInfoComponent> im
   }
 
   ngOnInit() {
-    this.eventsForm = this.createEventsForm();
-    if (this.event.contacts != null) {
-      this.event.contacts.forEach(x => {
-        this.contactsOrganizations.push({
-          id: x.id,
-          preName: x.preName,
-          name: x.name,
-          participated: false,
-          wasInvited: false,
-          modelType: MODEL_TYPE.CONTACT
-        });
-      });
-    }
-    if (this.event.organizations != null) {
-      this.event.organizations.forEach(x => {
-        this.contactsOrganizations.push({
-          id: x.id,
-          preName: x.name,
-          name: x.description,
-          participated: false,
-          wasInvited: false,
-          modelType: MODEL_TYPE.ORGANIZATION
-        });
-      });
-    }
-    if (this.event.participated != null) {
-      this.event.participated.forEach(x => {
-        let cont: ContactOrganizationDtoExtended = null;
-        if (x.modelType === MODEL_TYPE.CONTACT) {
-          cont = this.contactsOrganizations.find(y => y.modelType === MODEL_TYPE.CONTACT && y.id === x.objectId);
-        } else {
-          cont = this.contactsOrganizations.find(y => y.modelType === MODEL_TYPE.ORGANIZATION && y.id === x.objectId);
-        }
-        if (cont != null) {
-          cont.participated = x.hasParticipated;
-          cont.wasInvited = x.wasInvited;
-        }
-      });
-    }
-    this.modService.getSortedListByTypeAndId(this.event.id, MODEL_TYPE.EVENT)
-      .pipe(map(mod => mod.data), map(mod => mod.filter(el => el.dataType !== DATA_TYPE.NONE).sort(this.getSortHistoryFunction)))
-      .subscribe(result => {
-        this.dataHistory = result;
-      });
+    this.eventsForm = this.fb.group({
+      name: [''],
+      date: [''],
+      time: [''],
+      duration: [''],
+      description: [''],
+      location: ['']
+    });
+
+    this.contactsOrganizations.concat(this.event.contacts?.map(x => {
+      return {
+        id: x.id,
+        preName: x.preName,
+        name: x.name,
+        participated: false,
+        wasInvited: false,
+        modelType: MODEL_TYPE.CONTACT
+      };
+    }));
+
+    this.contactsOrganizations.concat(this.event.organizations?.map(x => {
+      return {
+        id: x.id,
+        preName: x.name,
+        name: x.description,
+        participated: false,
+        wasInvited: false,
+        modelType: MODEL_TYPE.ORGANIZATION
+      };
+    }));
+
+    this.event.participated?.forEach(x => {
+      let cont: ContactOrganizationDtoExtended = null;
+      if (x.modelType === MODEL_TYPE.CONTACT) {
+        cont = this.contactsOrganizations.find(y => y.modelType === MODEL_TYPE.CONTACT && y.id === x.objectId);
+      } else {
+        cont = this.contactsOrganizations.find(y => y.modelType === MODEL_TYPE.ORGANIZATION && y.id === x.objectId);
+      }
+      if (cont != null) {
+        cont.participated = x.hasParticipated;
+        cont.wasInvited = x.wasInvited;
+      }
+    });
+    // Load initial modification entries
+    this.loadModifications(0, 5);
     this.eventsForm.patchValue(this.event);
     this.eventsForm.get('date').patchValue(this.formatDate(this.event.date));
     this.eventsForm.get('time').patchValue(this.formatTime(this.event.time));
@@ -126,12 +133,16 @@ export class EventsInfoComponent extends BaseDialogInput<EventsInfoComponent> im
     return [hours, minutes].join(':');
   }
 
-  private createEventsForm(): FormGroup {
-    return this.fb.group({
-      name: [''],
-      date: [''],
-      time: [''],
-      duration: ['']
-    });
+  onPaginationChangedModification(event: PageEvent) {
+    this.loadModifications((event.pageIndex * event.pageSize), event.pageSize);
+  }
+
+  private loadModifications(pageStart: number, pageSize: number) {
+    this.modService.getSortedListByTypeAndId(this.event.id, MODEL_TYPE.EVENT, pageStart, pageSize)
+      .subscribe(result => {
+        this.dataHistory = result.data;
+        this.modificationsPaginationLength = result.totalRecords;
+      });
+
   }
 }
