@@ -34,9 +34,21 @@ export class OrganizationsListComponent implements OnInit, OnDestroy {
 	length = 0;
 	isAdminUserLoggedIn = false;
 	dataSource = new MatTableDataSource<OrganizationDto>();
+	permissionAdd = false;
+	permissionModify = false;
+	permissionDelete = false;
+	permissionAddHistory = false;
 
-	constructor(public dialog: MatDialog, service: OrganizationService,
-		           private changeDetectorRefs: ChangeDetectorRef, private mediaObserver: MediaObserver, private jwt: JwtService) {
+	selectedRow = 0;
+	selectedCheckBoxList: Array<number> = new Array<number>();
+	isAllSelected = false;
+
+	constructor(
+		public dialog: MatDialog,
+		service: OrganizationService,
+		private changeDetectorRefs: ChangeDetectorRef,
+		private mediaObserver: MediaObserver,
+		private jwt: JwtService) {
 		this.service = service;
 		this.flexMediaWatcher = mediaObserver.asObservable().subscribe((change: MediaChange[]) => {
 			if (change[0].mqAlias !== this.currentScreenWidth) {
@@ -50,32 +62,36 @@ export class OrganizationsListComponent implements OnInit, OnDestroy {
 		const filterValue = (event.target as HTMLInputElement).value;
 		this.dataSource.filter = filterValue.trim().toLowerCase();
 		this.dataSource.filterPredicate = ((data, filter) => {
-		  if (data.name.trim().toLowerCase().includes(filter) || data.description.trim().toLowerCase().includes(filter) ||
-			data.address.street.trim().toLowerCase().includes(filter) || data.address.streetNumber.trim().toLowerCase().includes(filter) ||
-			data.address.zipcode.trim().toLowerCase().includes(filter) || data.address.city.trim().toLowerCase().includes(filter) ||
-			data.address.country.trim().toLowerCase().includes(filter) || data.contact.phoneNumber.trim().toLowerCase().includes(filter) ||
-			data.contact.mail.trim().toLowerCase().includes(filter) || data.contact.fax.trim().toLowerCase().includes(filter)) {
-			return true;
-		  } else {
-			return false;
-		  }
+			if (data.name.trim().toLowerCase().includes(filter) || data.description.trim().toLowerCase().includes(filter) ||
+				data.address.street.trim().toLowerCase().includes(filter) || data.address.streetNumber.trim().toLowerCase().includes(filter) ||
+				data.address.zipcode.trim().toLowerCase().includes(filter) || data.address.city.trim().toLowerCase().includes(filter) ||
+				data.address.country.trim().toLowerCase().includes(filter) || data.contact.phoneNumber.trim().toLowerCase().includes(filter) ||
+				data.contact.mail.trim().toLowerCase().includes(filter) || data.contact.fax.trim().toLowerCase().includes(filter)) {
+				return true;
+			} else {
+				return false;
+			}
 		});
-	  }
+	}
 
 	ngOnInit(): void {
 		this.isAdminUserLoggedIn = this.jwt.getUserId() === 1;
 		this.tagsFilter.setRefreshTableFunction(() => this.applyTagFilter());
 		this.getData();
+		this.permissionAdd = this.jwt.hasPermission('Anlegen einer Organisation');
+		this.permissionDelete = this.jwt.hasPermission('Löschen einer Organisation');
+		this.permissionModify = this.jwt.hasPermission('Einsehen und Bearbeiten aller Organisationen');
+		this.permissionAddHistory = this.jwt.hasPermission('Hinzufügen eines Historieneintrags bei Kontakt oder Organisation');
 	}
 
 	applyTagFilter() {
 		this.dataSource = new MatTableDataSource<OrganizationDto>();
 		this.allOrganizations.forEach(x => {
-		  if (this.tagsFilter.areAllTagsIncluded(x.tags)) {
-			this.dataSource.data.push(x);
-		  }
+			if (this.tagsFilter.areAllTagsIncluded(x.tags)) {
+				this.dataSource.data.push(x);
+			}
 		});
-	  }
+	}
 
 	ngOnDestroy(): void {
 		this.flexMediaWatcher.unsubscribe();
@@ -87,6 +103,7 @@ export class OrganizationsListComponent implements OnInit, OnDestroy {
 			this.displayedColumns = ['Name', 'Zugehörige', 'Action'];
 		} else {
 			this.displayedColumns = [
+				'Icon',
 				'Name',
 				'Beschreibung',
 				'Strasse',
@@ -106,20 +123,22 @@ export class OrganizationsListComponent implements OnInit, OnDestroy {
 	private getData() {
 		this.organizations = this.service.get();
 		this.organizations.subscribe(x => {
-			 this.length = x.length;
-			 this.dataSource.data = x;
-			 this.allOrganizations = x;
-			 this.tagsFilter.updateTagsInAutofill(this.allOrganizations);
-			 this.applyTagFilter();
+			this.length = x.length;
+			this.dataSource.data = x;
+			this.allOrganizations = x;
+			this.tagsFilter.updateTagsInAutofill(this.allOrganizations);
+			this.applyTagFilter();
+			this.selectedCheckBoxList = new Array<number>();
+			this.selectedRow = 0;
+			this.isAllSelected = false;
 		});
 		this.changeDetectorRefs.detectChanges();
-		// this.organizationMock = this.orgaMock.getOrganizationsMock();
 	}
 
 	openAddDialog() {
 		console.log('openedAddDialog');
 		const dialogRef = this.dialog.open(OrganizationsAddDialogComponent, {
-			disableClose: true
+			disableClose: true, height: '600px'
 		});
 		dialogRef.afterClosed().subscribe((result) => {
 			this.getData();
@@ -149,15 +168,15 @@ export class OrganizationsListComponent implements OnInit, OnDestroy {
 				this.service.postHistoryElement(y, id).subscribe(x => this.getData());
 			}
 		});
-	  }
+	}
 
 	openInfo(id: number) {
 		this.service.getById(id).subscribe(x => this.dialog.open(OrganizationsInfoComponent, { data: x }));
-	  }
+	}
 
 	deleteOrganization(id: number) {
 		const deleteDialogRef = this.dialog.open(DeleteEntryDialogComponent, {
-			data: 'Organisation',
+			data: 'organization.organization',
 			disableClose: true
 		});
 
@@ -170,22 +189,50 @@ export class OrganizationsListComponent implements OnInit, OnDestroy {
 
 	addDummyOrganization() {
 		this.service.post({
-		  name: 'Organisation' + this.length,
-		  description: 'Bezeichnung' + this.length,
-		  employees: [],
-		  address: {
-			city: 'Statd',
-			country: 'Land',
-			street: 'Strasse',
-			streetNumber: this.length.toString(),
-			zipcode: '12345'
-		  },
-		  contact: {
-			contactEntries: [],
-			fax: '0234234-234' + this.length,
-			mail: 'info@testorga' + this.length + '.de',
-			phoneNumber: '02342-234234' + this.length
-		  }
+			name: 'Organisation' + this.length,
+			description: 'Bezeichnung' + this.length,
+			employees: [],
+			address: {
+				city: 'Statd',
+				country: 'Land',
+				street: 'Strasse',
+				streetNumber: this.length.toString(),
+				zipcode: '12345'
+			},
+			contact: {
+				contactEntries: [],
+				fax: '0234234-234' + this.length,
+				mail: 'info@testorga' + this.length + '.de',
+				phoneNumber: '02342-234234' + this.length
+			}
 		}).subscribe(x => this.getData());
+	}
+
+	mouseOver(id: number) {
+		this.selectedRow = id;
+	}
+
+	isSelectedRow(id: number): boolean {
+		const selectedIndex = this.selectedCheckBoxList.find(a => a === id);
+		return this.selectedRow === id || selectedIndex != null;
+	}
+	onCheckBoxChecked(id: number) {
+		const position = this.selectedCheckBoxList.indexOf(id);
+		if (position > -1) {
+			this.selectedCheckBoxList.splice(position, 1);
+		} else {
+			this.selectedCheckBoxList.push(id);
+		}
+	}
+	changeSelectionAll() {
+		this.isAllSelected = !this.isAllSelected;
+		this.selectedCheckBoxList = new Array<number>();
+		if (this.isAllSelected) {
+			this.allOrganizations.forEach(x => this.selectedCheckBoxList.push(x.id));
+		}
+	}
+
+	isSelectionChecked(id: number) {
+		return this.selectedCheckBoxList.find(x => x === id) != null;
 	}
 }
