@@ -1,7 +1,7 @@
-import { OnInit, Component } from '@angular/core';
-import { MatDialogRef, MatDialog } from '@angular/material/dialog';
+import { OnInit, Component, Inject } from '@angular/core';
+import { MatDialogRef, MatDialog, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { HistoryElementCreateDto, HistoryElementType } from 'src/app/shared/api-generated/api-generated';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
 import { BaseDialogInput } from '../form/base-dialog-form/base-dialog.component';
 
 @Component({
@@ -11,22 +11,52 @@ import { BaseDialogInput } from '../form/base-dialog-form/base-dialog.component'
 })
 
 export class AddHistoryComponent extends BaseDialogInput<AddHistoryComponent> implements OnInit {
-    public types: string[] = ['E-Mail', 'Telefonanruf', 'Notiz', 'Besuch'];
+    public types = [
+        {
+            translate: 'contact.mail',
+            type: HistoryElementType.MAIL
+        },
+        {
+            translate: 'common.phoneCall',
+            type: HistoryElementType.PHONE_CALL
+        },
+        {
+            translate: 'common.note',
+            type: HistoryElementType.NOTE
+        },
+        {
+            translate: 'common.visit',
+            type: HistoryElementType.VISIT
+        }];
     public oppoSuitsForm: FormGroup;
     constructor(
         public dialogRef: MatDialogRef<AddHistoryComponent>,
         public dialog: MatDialog,
+        @Inject(MAT_DIALOG_DATA) public phoneNumber: string,
         private fb: FormBuilder) {
         super(dialogRef, dialog);
+        this.dialogRef.backdropClick().subscribe(() => {
+            // Close the dialog
+            dialogRef.close();
+        });
     }
 
     ngOnInit(): void {
+        const dateToInsert = this.getDateOfValueAsText(new Date(Date.now()));
         this.oppoSuitsForm = this.fb.group({
             type: [this.types[0], Validators.required],
-            date: ['', Validators.required],
+            date: [new FormControl(new Date(dateToInsert)), Validators.required],
             information: ['', Validators.required],
             comment: ['', Validators.required]
         });
+
+        if (this.phoneNumber != null && this.phoneNumber.length > 0) {
+            this.oppoSuitsForm.get('type').setValue(this.types[1]);
+            this.oppoSuitsForm.get('comment').setValue(this.phoneNumber);
+        }
+
+        this.oppoSuitsForm.get('date').setValue(dateToInsert);
+
     }
 
     hasChanged() {
@@ -35,22 +65,27 @@ export class AddHistoryComponent extends BaseDialogInput<AddHistoryComponent> im
 
     getObject(): HistoryElementCreateDto {
         let historyToSave: HistoryElementCreateDto;
-        let typeToSave = HistoryElementType.MAIL;
-        const typeText: string = this.oppoSuitsForm.get('type').value;
-        if (typeText === this.types[1]) {
-            typeToSave = HistoryElementType.PHONE_CALL;
-        } else if (typeText === this.types[2]) {
-            typeToSave = HistoryElementType.NOTE;
-        } else if (typeText === this.types[3]) {
-            typeToSave = HistoryElementType.VISIT;
-        }
+        const date = new Date(this.oppoSuitsForm.get('date').value);
         historyToSave = {
-            date: this.oppoSuitsForm.get('date').value,
+            date: this.getDateOfValueAsText(date),
             name: this.oppoSuitsForm.get('information').value,
-            type: typeToSave,
+            type: this.oppoSuitsForm.get('type').value,
             comment: this.oppoSuitsForm.get('comment').value,
         };
         return historyToSave;
+    }
+
+    getDateOfValueAsText(date: Date): string {
+        let day = date.getDate().toString();
+        if (day.length === 1) {
+            day = '0' + day;
+        }
+        let month = (date.getMonth() + 1).toString();
+        if (month.length === 1) {
+            month = '0' + month;
+        }
+        const text = date.getFullYear().toString() + '-' + month + '-' + day + 'T00:00:00';
+        return text;
     }
 
     public onCancel(): void {
