@@ -10,7 +10,12 @@ export class DeleteEntryDialogComponent implements OnInit {
 
   constructor(
     public dialogRef: MatDialogRef<DeleteEntryDialogComponent, IDeleteDialogReturnValue>,
-    @Inject(MAT_DIALOG_DATA) public data: string) { }
+    @Inject(MAT_DIALOG_DATA) public data: string) {
+        this.dialogRef.backdropClick().subscribe(() => {
+        // Close the dialog
+        dialogRef.close();
+      });
+    }
 
   public ngOnInit(): void {
   }
