@@ -27,28 +27,28 @@ describe('Contacts Tests', () => {
         cy.get('#preName').type('testVorname');
         cy.get('#contactPartner').type('razvan');
         cy.get('#gender')
-                .click()
-                .get('mat-option')
-                //get the female gender
-                .contains('Weiblich')
-                .click()  
+            .click()
+            .get('mat-option')
+            //get the female gender
+            .contains('Weiblich')
+            .click()
         cy.get('#street').type('testStrasse');
         cy.get('#streetNumber').type('33');
         cy.get('#zipcode').type('90478');
         cy.get('#city').type('testStadt');
         cy.get('#country')
-                .click()
-                .get('mat-option')
-                //get the germany as country
-                .contains('Deutschland')
-                .click()  
+            .click()
+            .get('mat-option')
+            //get the germany as country
+            .contains('Deutschland')
+            .click()
         cy.get('#mail').type('info@test.de');
         cy.get('#fax').type('0123-123');
         cy.get('#phoneNumber').type('014-234234');
-        cy.get('#save').click({force: true});
-        cy.wait(2000);
+        cy.get('#save').click({ force: true });
+        cy.wait(10000);
         cy.get('#contactsTable').should("contain.text", 'testName');
-    }); 
+    });
     it('should correctly edit the fields gender and contactPerson of an existing user', () => {
         // Login with credentials admin/@dm1n1stR4tOr
         doLogin('admin', '@dm1n1stR4tOr');
@@ -60,19 +60,16 @@ describe('Contacts Tests', () => {
         cy.get('.editButton').click();
         cy.wait(4000);
         cy.get('#gender')
-                .click()
-                .get('mat-option')
-                //get the male gender
-                .contains('Männlich')
-                .click()  
+            .click()
+            .get('mat-option')
+            //get the male gender
+            .contains('Männlich')
+            .click()
         cy.get('#contactPartner').type('matis');
         cy.get('#saveEdit').click();
-        cy.wait(2000);
-        cy.get('.infoButton').click();
+        cy.wait(3000);
+        cy.get('.editButton').click();
         cy.wait(4000);
-        cy.get('#gender').should("have.value", 'Männlich');
         cy.get('#contactPartner').should("have.value", 'razvanmatis');
     });
 });
-
-
